Time: 7 ms (82.48%) | Memory: 18 MB (19.65%) - LeetSync

+++ 167-two-sum-ii-input-array-is-sorted/two-sum-ii-input-array-is-sorted.cpp
@@ -6,7 +6,7 @@ public:
         int right = n-1;
         int left = 0;
         while(left < right){
-            long sum = numbers[left] + numbers[right];
+            int sum = numbers[left] + numbers[right];
             if(sum == target){
                 return {left+1, right+1};
             } else if (sum<target){
